refactor : Sections 일급컬렉션 클래스 추가, Line 클래스 필드 교체
